Make codebase compatible with Node.js v10
`Object.fromEntries()` is an ES2019 feature, which was introduced in Node.js v12.
It was the only v12-only feature in this codebase,
and was used as part of code that swapped the keys and values of an object.
This is a task that lodash (a dependency that we already have) can also perform with its `invert` function,
so replacing that use of `Object.fromEntries()` with `lodash.invert()`
allows this package to become compatible with Node.js v10.